+++ tests/django/test_middleware.py
@@ -2,6 +2,7 @@ from unittest import mock
 from unittest.mock import MagicMock
 
 from django.views import View
+from rest_framework.viewsets import ViewSet
 
 from touchstone.django import InjectViewsMiddleware
 
@@ -14,35 +15,40 @@ class MyCls(MyAbc):
     pass
 
 
-class MyView(View):
+class DjangoView(View):
     obj: MyAbc
 
     def get(self):
         return self.obj
 
 
+class DRFViewSet(ViewSet):
+    obj: MyAbc
+
+    def retrieve(self):
+        return self.obj
+
+
 class TestInjectViewsMiddleware:
     def test_process_view_django_style(self):
-        view_func = MyView.as_view()
+        view_func = DjangoView.as_view()
         with mock.patch(
-            "touchstone.django.middleware.inject_magic_properties"
+                "touchstone.django.middleware.inject_magic_properties"
         ) as mock_inject_magic_properties:
             middleware = InjectViewsMiddleware(MagicMock())
             request = None
             middleware.process_view(request, view_func, [], {})
 
-        mock_inject_magic_properties.assert_called_once_with(MyView)
+        mock_inject_magic_properties.assert_called_once_with(DjangoView)
 
     def test_process_view_drf_style(self):
-        view_func = MyView.as_view()
-        del view_func.view_class
-        view_func.cls = MyView
+        view_func = DRFViewSet.as_view({'get': 'retrieve'})
 
         with mock.patch(
-            "touchstone.django.middleware.inject_magic_properties"
+                "touchstone.django.middleware.inject_magic_properties"
         ) as mock_inject_magic_properties:
             middleware = InjectViewsMiddleware(MagicMock())
             request = None
             middleware.process_view(request, view_func, [], {})
 
-        mock_inject_magic_properties.assert_called_once_with(MyView)
+        mock_inject_magic_properties.assert_called_once_with(DRFViewSet)