+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.1
 **Bug Fixes*
-* Fix a but where the `InjectViewsMiddleware` Django middleware doesn't work
+* Fix a bug where the `InjectViewsMiddleware` Django middleware doesn't work
   on Django Rest Framework (DRF) style ViewSets.
 
 ## 2.0.0